+++ lib/pusher_client_socket.dart
@@ -1,61 +1,35 @@
 library pusher_client_socket;
 
+import 'dart:async';
 import 'dart:convert';
+import 'dart:io';
 
-import 'package:pusher_client_socket/channel/channel.dart';
+import 'package:pusher_client_socket/channels/channel.dart';
 import 'package:pusher_client_socket/src/channels.collection.dart';
 import 'package:pusher_client_socket/src/events_listeners.collection.dart';
 import 'package:web_socket_client/web_socket_client.dart';
 
-import 'src/auth_options.dart';
 import 'src/options.dart';
 
-export 'channel/channel.dart';
+export 'channels/channel.dart';
+export 'src/options.dart';
 export 'package:web_socket_client/src/connection_state.dart';
 
+/// Represents a client for connecting to a Pusher server.
 class PusherClient {
-  PusherOptions? options;
+  /// The options of the client.
+  final PusherOptions options;
 
-  PusherClient({
-    Protocol protocol = Protocol.ws,
-    String? host,
-    required String authEndpoint,
-    required String key,
-    String? cluster,
-    int activityTimeout = 120000,
-    int pongTimeout = 30000,
-    Map<String, String> authHeaders = const {},
-    Map<String, String> parameters = const {},
-    bool enableLogging = false,
-  }) {
-    options = PusherOptions(
-      protocol: protocol,
-      host: host,
-      key: key,
-      cluster: cluster,
-      activityTimeout: activityTimeout,
-      pongTimeout: pongTimeout,
-      parameters: {
-        "protocol": "7",
-        "client": "flutter",
-        "version": "0.0.1",
-        "flash": "false",
-        ...parameters,
-      },
-      authOptions: AuthOptions(
-        authEndpoint,
-        headers: {
-          // "Content-Type": "application/json",
-          "Accept": "application/json",
-          ...authHeaders,
-        },
-      ),
-      enableLogging: enableLogging,
-    );
+  PusherClient({required this.options}) {
+    onConnectionError(_onConnectionError);
     onConnectionEstablished(_onConnectionEstablished);
-    onError(_onError);
-    listen("pusher:ping", _onPing);
-    // listen("pusher:pong", _onPong);
+    bind("pusher:ping", _onPing);
+    bind("pusher:pong", (data) {
+      _stopActivityTimer();
+      _resetActivityCheck();
+    });
+
+    if (options.autoConnect) connect();
   }
 
   WebSocket? __socket;
@@ -66,19 +40,22 @@ class PusherClient {
 
   bool _connected = false;
 
+  /// Returns a boolean value indicating whether the client is connected.
   bool get connected => _connected;
 
   ConnectionState _connectionState = const Disconnected();
 
+  /// Returns the current connection state of the client.
   ConnectionState get connectionState => _connectionState;
 
   WebSocket get _socket {
     if (__socket != null) return __socket!;
-    throw Exception("the channel does not initialized yet");
+    throw Exception("The channel does not initialized yet");
   }
 
+  /// Connects the client to the server.
   void connect() {
-    __socket = WebSocket(Uri.parse(options!.url));
+    __socket = WebSocket(Uri.parse(options.url));
 
     _socket.connection.listen(_onConnectionStateChange);
 
@@ -86,6 +63,13 @@ class PusherClient {
       _onMessageReceived,
       onError: (err) => _onEvent('error', err),
     );
+  }
+
+  /// Disconnects the client from the server.
+  void disconnect([int? code, String? reason]) {
+    options.log("DISCONNECT", null, "code: $code\n  reason: $reason");
+
+    _socket.close(code, reason);
   }
 
   void _onConnectionStateChange(ConnectionState state) {
@@ -98,7 +82,7 @@ class PusherClient {
       const Disconnected(): 'DISCONNECTED',
     };
 
-    options!.log(
+    options.log(
       "CONNECTION_STATE_CHANGED",
       null,
       "the connection state changed from ${states[_connectionState]} to ${states[state]}",
@@ -114,29 +98,86 @@ class PusherClient {
         break;
       case Connected():
         _onEvent('connected', null);
-        // _resetActivityCheck();
+        _resetActivityCheck();
         break;
       case Reconnecting():
         _onEvent('reconnecting', null);
         break;
       case Reconnected():
         _onEvent('reconnected', null);
+        _resetActivityCheck();
         break;
       case Disconnecting():
         _onEvent('disconnecting', null);
         break;
       case Disconnected():
         _onEvent('disconnected', state);
-        if (state.error != null) _onError(state.error);
-
+        if (state.error != null) {
+          _onEvent("connection_error", state.error);
+        }
+        _stopActivityTimer();
+        _connected = false;
         _socketId = null;
+        __socket = null;
         break;
       default:
     }
   }
 
+  void _onConnectionError(dynamic error) {
+    options.log("CONNECTION_ERROR", null, "error: $error");
+
+    disconnect(1006, error.message);
+
+    if (error is WebSocketException) {
+      connect();
+    }
+  }
+
+  void _onConnectionEstablished(Map data) {
+    options.log("CONNECTION_ESTABLISHED", null, "data: $data");
+    _socketId = data['socket_id'];
+    _connected = true;
+    _reSubscribe();
+  }
+
+  void _onPing(data) {
+    options.log("PINGING", null, "data: $data");
+    sendEvent("pusher:pong", data);
+  }
+
+  Timer? _activityTimer;
+
+  void _sendActivityCheck() {
+    _stopActivityTimer();
+
+    sendEvent("pusher:ping", null);
+
+    _activityTimer = Timer.periodic(
+      Duration(milliseconds: options.pongTimeout),
+      (timer) {
+        _onEvent("pusher:error", "Activity timeout");
+        disconnect(null, "Activity timeout");
+      },
+    );
+  }
+
+  void _resetActivityCheck() {
+    _stopActivityTimer();
+    _activityTimer = Timer.periodic(
+      Duration(milliseconds: options.activityTimeout),
+      (timer) => _sendActivityCheck(),
+    );
+  }
+
+  void _stopActivityTimer() {
+    _activityTimer?.cancel();
+  }
+
+  final _eventsListeners = EventsListenersCollection();
+
   void _onMessageReceived(message) {
-    options!.log("MESSAGE_RECEIVED", null, "$message");
+    options.log("MESSAGE_RECEIVED", null, "$message");
 
     dynamic event;
 
@@ -180,46 +221,16 @@ class PusherClient {
     }
   }
 
-  void _onConnectionEstablished(Map data) {
-    options!.log("CONNECTION_ESTABLISHED", null, "data: $data");
-    _socketId = data['socket_id'];
-    _connected = true;
+  /// Binds a listener to an event.
+  void bind(String event, Function listener) {
+    options.log("EVENT_BINDING", "event: $event");
+
+    _eventsListeners.bind(event, listener);
   }
 
-  void _onError(error) {
-    options!.log("ERROR", null, "error: $error");
-  }
-
-  void _onPing(data) {
-    options!.log("PINGING", null, "data: $data");
-    sendEvent("pusher:pong", data);
-  }
-
-  final _eventsListeners = EventsListenersCollection();
-  final Map<String, Map<String, List<Function>>> _channelsEventsListeners = {};
-
-  void listen(String event, Function listener, [String? channel]) {
-    options!.log("EVENT_LISTENING", channel, "event: $event");
-
-    if (channel != null) {
-      if (!_channelsEventsListeners.containsKey(channel)) {
-        _channelsEventsListeners[channel] = {};
-      }
-
-      if (!_channelsEventsListeners[channel]!.containsKey(event)) {
-        _channelsEventsListeners[channel]![event] = [];
-      }
-
-      if (!_channelsEventsListeners[channel]![event]!.contains(listener)) {
-        _channelsEventsListeners[channel]![event]!.add(listener);
-      }
-    } else {
-      _eventsListeners.bind(event, listener);
-    }
-  }
-
+  /// Send an event to the server.
   void sendEvent(String event, [dynamic data, String? channel]) {
-    options!.log("SEND_EVENT", null, "event: $event\n  data: $data");
+    options.log("SEND_EVENT", null, "event: $event\n  data: $data");
     _socket.send(jsonEncode({
       "event": event,
       "data": data,
@@ -227,95 +238,86 @@ class PusherClient {
     }));
   }
 
+  /// Binding to the connection state change event.
   void onConnectionStateChange(Function(ConnectionState) listener) =>
-      listen("connection_state_changed", listener);
+      bind("connection_state_changed", listener);
 
-  void onConnecting(Function listener) => listen('connecting', listener);
+  /// Binding to the connection state change event.
+  void onConnecting(Function listener) => bind('connecting', listener);
 
-  void onConnected(Function listener) => listen('connected', listener);
+  /// Binding to the connection state change event.
+  void onConnected(Function listener) => bind('connected', listener);
 
+  /// Binding to the connection state change event.
   void onConnectionEstablished(Function listener) =>
-      listen("pusher:connection_established", listener);
+      bind("pusher:connection_established", listener);
 
-  void onReconnecting(Function listener) => listen('reconnecting', listener);
+  /// Binding to the connection state change event.
+  void onReconnecting(Function listener) => bind('reconnecting', listener);
 
-  void onReconnected(Function listener) => listen('reconnected', listener);
+  /// Binding to the connection state change event.
+  void onReconnected(Function listener) => bind('reconnected', listener);
 
-  void onDisconnecting(Function listener) => listen('disconnecting', listener);
+  /// Binding to the connection state change event.
+  void onDisconnecting(Function listener) => bind('disconnecting', listener);
 
-  void onDisconnected(Function listener) => listen('disconnected', listener);
+  /// Binding to the connection state change event.
+  void onDisconnected(Function listener) => bind('disconnected', listener);
 
-  void onError(Function(dynamic error) listener) => listen('error', listener);
+  /// Binding to the connection state change event.
+  void onConnectionError(Function(dynamic error) listener) =>
+      bind('connection_error', listener);
 
-  void disconnect([int? code, String? reason]) {
-    options!.log("DISCONNECT", null, "code: $code\n  reason: $reason");
-
-    _socket.close(code, reason);
-  }
+  /// Binding to the connection state change event.
+  void onError(Function(dynamic error) listener) =>
+      bind('pusher:error', listener);
 
   late final _channelsCollection = ChannelsCollection(this);
 
-  T channel<T extends Channel>(String channelName) =>
-      _channelsCollection.channel<T>(channelName);
+  /// Returns a channel by name.
+  T channel<T extends Channel>(String channelName, {bool subscribe = false}) =>
+      _channelsCollection.channel<T>(channelName, subscribe: subscribe);
 
-  PrivateChannel private(String channelName) => channel(
+  /// Returns a private channel by name.
+  PrivateChannel private(String channelName, {bool subscribe = false}) =>
+      channel(
         channelName.startsWith("private-")
             ? channelName
             : "private-$channelName",
+        subscribe: subscribe,
       );
 
-  PrivateChannel privateEncrypted(String channelName) => channel(
+  /// Returns a private encrypted channel by name.
+  PrivateChannel privateEncrypted(String channelName,
+          {bool subscribe = false}) =>
+      channel(
         channelName.startsWith("private-encrypted-")
             ? channelName
             : "private-encrypted-$channelName",
+        subscribe: subscribe,
       );
 
-  PresenceChannel presence(String channelName) => channel(
+  /// Returns a presence channel by name.
+  PresenceChannel presence(String channelName, {bool subscribe = false}) =>
+      channel(
         channelName.startsWith("presence-")
             ? channelName
             : "presence-$channelName",
+        subscribe: subscribe,
       );
 
-  T subscribe<T extends Channel>(String channelName) {
-    final channel = this.channel<T>(channelName);
-    channel.subscribe();
-    return channel;
-  }
+  /// Subscribes to a channel by name.
+  T subscribe<T extends Channel>(String channelName) =>
+      channel<T>(channelName, subscribe: true);
 
+  /// Unsubscribes from a channel by name.
   void unsubscribe(String channelName) => channel(channelName).unsubscribe();
 
-  // Timer? _activityTimer;
-
-  // void _sendActivityCheck() {
-  //   _stopActivityTimer();
-
-  //   sendEvent("pusher:ping", null);
-
-  //   _activityTimer = Timer.periodic(
-  //     Duration(milliseconds: options!.pongTimeout),
-  //     (timer) {
-  //       _onError("Activity timeout");
-  //       disconnect(408, "Activity timeout");
-  //     },
-  //   );
-  // }
-
-  // void _resetActivityCheck() {
-  //   _stopActivityTimer();
-  //   _activityTimer = Timer.periodic(
-  //     Duration(milliseconds: options!.activityTimeout),
-  //     (timer) {
-  //       _sendActivityCheck();
-  //     },
-  //   );
-  // }
-
-  // void _stopActivityTimer() {
-  //   _activityTimer?.cancel();
-  // }
-
-  // void _onPong() {
-  //   _stopActivityTimer();
-  //   _resetActivityCheck();
-  // }
+  void _reSubscribe() {
+    _channelsCollection.forEach((channel) {
+      if (channel.subscribed) {
+        channel.subscribe(true);
+      }
+    });
+  }
 }